requirements.txt
openai
unstructured
pdfplumber
python-dotenv
pydantic
tqdm
pandas